Enable undo/redo of iss and collapse.
Change ClintScop::transform() function so that it manages creation and
deletion of statement occurrences that was previously done in the
manipulation manager leaking the abstraction.

Transformation may throw, therefore the group should be present before
even if it is empty to maintain the undo/redo/guessing.  Make sure the
reference to the back of the complementary sequence gets copied when
necessary.